[java] Add tests related to #6146 - unknown constraint is compatible with other constraints
To fix #6146 we still need to fix the underlying issue,
which is that a resolution failure (ie, the `filter(...)`)
call here fails resolving after having tried out the overload,
and that doesn't clean up the whole tree from inference variables.

As a last step we can check that the cast is feasible in TypeTestUtil,
although it will brush future issues under the rug.