Fix problem with coordinate translation
The problem is that CPD needs to store input
coordinates, and it doesn't have access to the
translated TextDocument that is created by the
CharStream in our Javacc wrapper code. This means
CPD needs cooperation from the CpdLexer
implementations to not mess up the coordinate system,
and it is difficult to check for mixups.

This is only a problem in CpdLexers that both
1. Use the offset coordinates (which are new in this branch so aren't
   used by anyone yet), and
2. Perform some translation on the TextDocument, eg to interpret
   escapes.

Some lexers meet the second criterion and none meet the first,
so this is not yet a problem. But, migrating lexers to the new
offset coordinate mode needs to be aware of this.