Make CPD SimpleRenderer show starting point of dupe
The reason I add this is that it is frequent for matches to
start on the _previous_ relevant line, because the previous
line is also terminated by the same kind of token. Eg commas,
semicolons, or closing braces are commonly found at the end
of lines, and will count towards the match, even though the
"relevant" content is on the next line.

For instance for the following text:

	var x = [
	  1, 1, 1, 1, 1, 1, 1, 1,
	  0, 0, 0, 0, 0, 0, 0, 0,
	  2, 2, 2, 2, 2, 2, 2, 2,
	  0, 0, 0, 0, 0, 0, 0, 0,
	  3, 3, 3, 3, 3, 3, 3, 3,
	  0, 0, 0, 0, 0, 0, 0, 0,
	  4, 4, 4, 4, 4, 4, 4, 4
	];

The output is now

	Found a 2 line (2 tokens) duplication in the following files:
	Starting at line 2 of /var/Foo.java
	Starting at line 4 of /var/Foo.java
	Starting at line 6 of /var/Foo.java

	------------------------v starting from here (col 25)
	  1, 1, 1, 1, 1, 1, 1, 1,
	  0, 0, 0, 0, 0, 0, 0, 0,
	------------------------^ ending here (col 25)

Before it was

	Found a 2 line (2 tokens) duplication in the following files:
	Starting at line 2 of /var/Foo.java
	Starting at line 4 of /var/Foo.java
	Starting at line 6 of /var/Foo.java

	  1, 1, 1, 1, 1, 1, 1, 1,
	  0, 0, 0, 0, 0, 0, 0, 0,

which was misleading because it suggests the full 2 lines are
duplicated, when only the sequence `, 0, 0, ...` (the second
line and the last token of the first line) is.