[ci] Exclude build resources from spring-framework for regression tester
Added an exclude pattern for build resources.

Those resources are copied from the source tree. The regression tester tries to link to those when violations occur, but those are not available in the github repo. The violations are anyway duplicates because they are also reported correctly on the file copy that is in the source tree.