Fix #4953 - deprecate methods of PMDConfiguration that use ClassLoader
Add new methods that just set/get the string classpath.
ClasspathClassLoader is now created not in PMD core but
within JavaLanguageProcessor.

One of the main differences with main is that we accept
file:// URLs as valid classpath entries. They may refer
to directories or to jar files. This is consistent with
how java.net.URLClassLoader treats classpath entries.
If you want to use a classpath file (listing classpath
entries one per line) you need to use the corresponding,
separate method in PMDConfiguration.