Introduce explicit wrapper
This clears up possible confusions and enables new use cases.
For instance, we need to allow running different analyses with
the same classpath object (classloader). This is crucial eg in
tests, where we cannot afford to create thousands of identical
classloaders.

The new API is more explicit about whether the classloaders are
closed by PMD or are expected to be closed by the user.

This removes a risk of memory leak caused by calling prependAuxClasspath
repeatedly. Previously this would have created several
ClasspathClassLoader, but only the last one would be closed.

TODO: are there still risks of memory leaks?
- Calling setClasspathWrapper moves the previous CPW out of scope.
This is a potential memory leak -> it is likely we should close it
and log it. We expect people to call this method with their own
custom classloader. Technically they could also do

config.setClasspathWrapper(config.getClasspathWrapper())

so we cannot close the CP wrapper early. We actually cannot build it
early. The prependclasspath should not create any closeable resources.