Make PmdClasspathWrapper not Closeable and immutable
Ref #4953

This is a huge simplification and also fixes many of the
usability issues with the previous iteration. For instance
it doesn't matter if the classpath wrapper goes out of scope
because it doesn't actually manage any resources.

Also if you think about it you don't need a real chain of
classloaders with a boolean to know if you can close it or
not. You just need a fallback classloader and the classpath
as a string (or even better, a pre-parsed list of URLs).

This is really the "graceful collapse of complexity" Brian Goetz
is talking about.